// src/app/core/enums/login-services.enum.ts
export enum LoginServicesEnum {
  LOGIN = 'login',
  GET_USER_BY_EMAIL = 'getUserByEmail',
  GET_USER_BY_PHONE = 'getUserByPhone',
  RESTORE_PWD = 'restorePwd',
  SEND_SMS = 'sendSms',
  SEND_WHATSAPP = 'sendWhatsapp',
  SEND_EMAIL = 'sendEmail',
  RESEND_EMAIL_CODE = 'reSendEmailCode',
  RETURN_HOME = 'returnHome',
  CONFIRM_USER = 'confirmUser',
  RESEND_CONFIRM_CODE = 'reSendConfirmCode',
}